requirements.txt
python-telegram-bot
python-dotenv
cryptography
openpyxl
fastapi
uvicorn
psycopg2-binary
yt-dlp